Consolidate logic
